refactor(core): reuse workspace destroy contract
